+++ solidity/dashboard/src/components/AddETHModal.jsx
@@ -1,10 +1,7 @@
 import React, { useCallback } from "react"
-import { SubmitButton } from "./Button"
-import { useCustomOnSubmitFormik } from "../hooks/useCustomOnSubmitFormik"
-import FormInput from "./FormInput"
+import AvailableETHForm from "./AvailableETHForm"
 import { getErrorsObj } from "../forms/common-validators"
 import { withFormik } from "formik"
-import { colors } from "../constants/colors"
 import web3Utils from "web3-utils"
 import { useWeb3Context } from "./WithWeb3Context"
 import { tbtcAuthorizationService } from "../services/tbtc-authorization.service"
@@ -53,6 +50,7 @@ const AddEthModal = ({ operatorAddress, closeModal }) => {
         onSubmit={onSubmit}
         yourAddress={yourAddress}
         closeModal={closeModal}
+        action="add"
       />
     </>
   )
@@ -60,42 +58,6 @@ const AddEthModal = ({ operatorAddress, closeModal }) => {
 
 export default React.memo(AddEthModal)
 
-const AddETHForm = ({ onSubmit, closeModal, ...formikProps }) => {
-  const onSubmitBtn = useCustomOnSubmitFormik(onSubmit)
-
-  return (
-    <form>
-      <FormInput
-        name="ethAmount"
-        type="text"
-        label="ETH Amount"
-        placeholder="0"
-      />
-      <div
-        className="flex row center mt-2"
-        style={{
-          borderTop: `1px solid ${colors.grey20}`,
-          margin: "0 -2rem",
-          padding: "2rem 2rem 0",
-        }}
-      >
-        <SubmitButton
-          className="btn btn-primary"
-          type="submit"
-          onSubmitAction={onSubmitBtn}
-          withMessageActionIsPending={false}
-          triggerManuallyFetch={true}
-          disabled={!formikProps.dirty}
-        >
-          add eth
-        </SubmitButton>
-        <span onClick={closeModal} className="ml-1 text-link">
-          Cancel
-        </span>
-      </div>
-    </form>
-  )
-}
 const AddETHFormik = withFormik({
   validateOnChange: false,
   validateOnBlur: false,
@@ -127,4 +89,4 @@ const AddETHFormik = withFormik({
     })
   },
   displayName: "AddEthForm",
-})(AddETHForm)
+})(AvailableETHForm)
