+++ solidity/dashboard/src/components/WithdrawETHModal.jsx
@@ -1,16 +1,13 @@
 import React, { useCallback } from "react"
-import { SubmitButton } from "./Button"
-import { useCustomOnSubmitFormik } from "../hooks/useCustomOnSubmitFormik"
-import FormInput from "./FormInput"
 import { getErrorsObj } from "../forms/common-validators"
 import { withFormik } from "formik"
-import { colors } from "../constants/colors"
 import web3Utils from "web3-utils"
 import { useWeb3Context } from "./WithWeb3Context"
 import { tbtcAuthorizationService } from "../services/tbtc-authorization.service"
 import { useShowMessage, messageType } from "./Message"
 import * as Icons from "./Icons"
 import AvailableEthAmount from "./AvailableEthAmount"
+import AvailableETHForm from "./AvailableETHForm"
 
 const WithdrawETHModal = ({ operatorAddress, availableETH, closeModal }) => {
   const web3Context = useWeb3Context()
@@ -60,6 +57,7 @@ const WithdrawETHModal = ({ operatorAddress, availableETH, closeModal }) => {
         onSubmit={onSubmit}
         availableETH={availableETH}
         closeModal={closeModal}
+        action="withdraw"
       />
     </>
   )
@@ -67,42 +65,6 @@ const WithdrawETHModal = ({ operatorAddress, availableETH, closeModal }) => {
 
 export default React.memo(WithdrawETHModal)
 
-const WithdrawETHForm = ({ onSubmit, closeModal, ...formikProps }) => {
-  const onSubmitBtn = useCustomOnSubmitFormik(onSubmit)
-
-  return (
-    <form>
-      <FormInput
-        name="ethToWithdraw"
-        type="text"
-        label="ETH Amount"
-        placeholder="0"
-      />
-      <div
-        className="flex row center mt-2"
-        style={{
-          borderTop: `1px solid ${colors.grey20}`,
-          margin: "0 -2rem",
-          padding: "2rem 2rem 0",
-        }}
-      >
-        <SubmitButton
-          className="btn btn-primary"
-          type="submit"
-          onSubmitAction={onSubmitBtn}
-          withMessageActionIsPending={false}
-          triggerManuallyFetch={true}
-          disabled={!formikProps.dirty}
-        >
-          withdraw eth
-        </SubmitButton>
-        <span onClick={closeModal} className="ml-1 text-link">
-          Cancel
-        </span>
-      </div>
-    </form>
-  )
-}
 const WithdrawETHFormik = withFormik({
   validateOnChange: false,
   validateOnBlur: false,
@@ -136,4 +98,4 @@ const WithdrawETHFormik = withFormik({
     return getErrorsObj(errors)
   },
   displayName: "WithdrawETHForm",
-})(WithdrawETHForm)
+})(AvailableETHForm)
